fix login fail issue
